x86-64: Create a BOA constructor named EA for instances of EA
And make SIZE an optional argument defaulting to :UNSPECIFIED.

This is step 1 of multiple changes needed to avoid messing up EQness
of register operand TNs, as well as operate on stack TNs in sizes other
than :QWORD.  The problem is that EAs and stack TNs have to be fungible;
and those, as well as registers, must allow operating on sub-words
without need for REG-IN-SIZE causing loss of identity of IR.

To begin with, EAs have to be sizeless, because creating a new EA to
process a byte of some stack TN is harmful to semantics. And sizes larger
than :QWORD were usually wrong anyway. In addition, 36% of our uses of EAs
were for LEA which ignores the source size. Not to mention, CMPXCHG16B
and maybe a few others that should care didn't care.

So most, but not all, EAs have the size removed in this change. The size
is implicit except for instructions that move an immediate value
to memory or perform an ALU op like ADD to memory with an immediate.
The next steps will be to:
* Add mandatory size qualifiers to MOVSX, MOVZX and some other
  instructions that have differently sized source/destination.
* Add optional size qualifiers to operations that could use them,
  such as (INST CMP :BYTE SOME-TN SOME-WIDETAG) etc.

This imparts no functional change despite the massive diff, because opcodes
taking two operands of the same size call MATCHING-OPERAND-SIZE which would
have already been failing had we passed in an EA whose size was unexpected.
Thus, changing any size to :UNSPECIFIED in the EA behaves as before: match
whatever the other operand is. Additionally the change was machine-verified
by the file in tools-for-build which can be used on the 32-bit code too.