Simplify :USING disassembler printer spec
The complication stemmed from an older technique for coalescing printer
lambdas so that only one would be compiled for any given s-expressions that
matched on some similarity test that is almost like EQUAL but different.
I don't understand this code any more, but removing the junk didn't seem
to harm the x86-64 backend which relies on a :USING clause.