Optimize ALIEN-FUNCALL-SAVES-FP-AND-PC.
Instead of saving the caller PC and FP save the current frame pointer,
from which PC and FP can be fetched. That allows to save a single
fixnum instead of a DX-allocated array.
Do to not link to the previous saved FP using a cons, just bind the
variable and the previous values can be retrieved from the binding
stack.