Add optional size modifier to a bunch of x86-64 instructions
In the end I did not the design that gave DEFINE-INSTRUCTION a purely
declarative syntax for this rather than requiring shuffling the operands
by hand. It would have handled the CMPx family of instructions for PPC
which has a similiar issue for the optional condition register field.
But it didn't seem terribly important, so this is ad-hoc like everything else.

There is no change to how the assembler is used yet.