Remove 'search' command from ldb
Stepping through the heap 2 words at a time as it did, rather than by object
it as likely to construct an illegal pointer (and crash) as not.
This was particularly a problem with interiors of code objects, unboxed arrays,
and instances with raw slots. In addition, the wrong lowtag was reconstructed
for a closure, instance, or funcallable-instance.
It might have been possible to make the command correct or safe, or both,
but bractically speaking, any use of 'search' could be served by the
'-d' ("decode") option of the 'dump' command and a visual scan.