x86-64/linux: move static space higher and enlarge immobile spaces
This plays better with any combination of tcmalloc and clang sanitizers.
The immobile space size issue is orthogonal to the small space start,
and it merely raises the immobile space consumption to be approx +15% of the
default dynamic space size versus +12% of same.

We've been using this combination of settings for over a year and a half.