Implement ICF more correctly
which means: as correctly as can be given an assumption that end users
will not be harmed by function pointers that become EQ due to folding.

Each backend will need an equivalence predicate to decide whether code
objects are functionally the same. Only x86-64 has that predicate for now,
and the other backends disable ICF.

The weak level of folding has been made stronger (will fold more).
And since this behavior could still be considered surprising, it is only
done for self-build. The x86-64 image shrinks by about 200K.