Fix a possible type error in genesis from rev 04139ea0e0f83a13
A (VECTOR (SIMPLE-ARRAY (UNSIGNED-BYTE 8) (#.SB-VM:N-WORD-BYTES))) was assigned
into the OUTER-VECTOR slot of a BIGVEC. It's compatible with
  (upgraded-array-element-type '(simple-array (unsigned-byte 8) (65536)))
but not compatible with a declared type of that.

SBCL 1.0.55 considers the code to be wrong. I suspect that's right, which could
either mean we have a regression, or that we deliberately weakened a check.