x86-64: defrag fdefns, GFs, and closure trampolines into varyobj space
This is done only when saving a core since the runtime allocator still uses
fixed-size object pages for these three object types, for simplicity and speed.
Transporting to variable-sized object pages allows assigning C symbols
to their addresses when processed with 'editcore'.

The garbage-collector required a few minor changes to allow fdefns and
funcallable instances on either variable-sized or fixed-size object pages.
Specifically, the variable-sized pages use a software write barier (sans
protection), so we have to deal with that when assigning slots. As was the
case before, this is not a great implementation of a software write barrier,
but it works. (I plan to implement a better one). The impact of an extra
asm routine call when assigning to fdefns is hardly significant. Similary
the CHANGE-CLASS protocol on a generic function is hundreds if not thousands
of function calls, so one extra is not even noticeable.