Mostly get PIE support working with ELFinated cores
* Remove assumptions from immobile-space GC that pointers are only 32 bits.
  Similarly in in traceroot.

* Add assertions that coreparse won't trash your code when relocating.

* Emit explicit relocations from editcore so that coreparse avoids walking
  the whole Lisp heap.

The remaining problem are twofold:
(1) There is always going to be a need to perform fixups at startup,
    but such is generally the case with PIE - the system loader performs
    relocations for C code, but Lisp tends to contain more absolute pointers
    to code than C code contains. This is not a show-stopper, just annoying.
(2) any new fdefns allocated by COMPILE need to go into the varyobj space
    which the allocator can't do that yet. This is a show-stopper.

Also a small secret: for x86-64 linux with :SB-DYNAMIC-CORE, non-ELF cores
already support the C runtime being compiled with '-pie'.
The reason for avoiding ASLR and PIE on 32-bit systems was that randomizing
might conflict with Lisp's static/read-only, and/or linkage table spaces,
and there's nothing that can be done to remedy that.
PIE x86-64 places text in the stratosphere as far as address space goes.