Add GC support for lockfree singly-linked lists
Based on paper by Tim Harris @ https://timharris.uk/papers/2001-disc.pdf.
The algorithm depends on arbitrarily manipulation of 1 bit of a pointer
which we can do by relaxing the tagging requirement for pointers from a
node to its successor. For technical reasons, our representation differs
from that in the reference paper - for us, the lowtag bits of a successor
pointer are either all correct (INSTANCE_POINTER_LOWTAG), or all 0.

The code has been lightly exercised on arm64, ppc, x86, and x86-64.
I'm still deciding on the interface that we want to expose,
and where such lists can be used internally.