pacify a lockfree list test
expecting a performance gain from threading on a system with a
small number of cores is unlikely to lead to happiness.  Use
cut-and-paste programming to modify the test to expect modest
gains with high core count and no gains with low core count.

(OAOOM programming since... a long time ago)