fix lp#1804796
Basically, MAKE-NUMERIC-TYPE needs to return the canonical REAL/FLOAT
types, not ersatz variants, if that is what it is being asked to do.
Mostly it is not asked to do this, but in one or two places in the
system, notably the COMPLEX DERIVE-TYPE optimizer, it does.

In the process, repurpose COERCE-NUMERIC-BOUND to help in making
numeric types, and invent BOUNDS-UNBOUNDED-P to capture an idiom
useful in the type system.