Fix example for fix to map-allocated-objects
In the example as it was, the scan would have started at the right place,
making it a non-example. In the revised example, the scan would have started
at the wrong place if consing happened as described. And there are probabl
more edge cases, especially if threads are involved, though this particular
failure was observed in single-threaded code.