Change placement of simple fun offset table by a tiny bit
This change aligns the simple-fun-offset table to exactly abut the end of its
object, as will be needed for a subsequent change in which we have code->header
store the exact object length in words. The count of boxed words will go in the
next word, and the unboxed byte count is not explicitly stored.
The text size can be recovered from the start of the fun offset table.
We just need to know how many boxed words to subtract out, and possibly a small
padding amount which is already stored to make the x86 disassembler happier.