Remove a few dubious 'movable' annotations
These pertain to code motion optimizations, but are in fact completely useless,
since the compiler literally makes no use of them. However, they should at least
not be misleading. Querying an object's GC page or address is not a candidate
for code motion. Similarly, in (%bignum-set-length b n) ... (%bignum-length b)
the latter is not a candidate to move back in time.