Implement a faithful model of the target CL package for the host
The function SB-XC:SYMBOL-PACKAGE returns *CL-PACKAGE* when and only when
it ought to, so that we needn't worry about host design choices that export
symbols from CL whose home packages are elsewhere
such as (symbol-package 'find-class) => #<PACKAGE CLOS> in CLISP.

This change introduces a new idiom for bug avoidance: if the behavior differs
in such a subtle way that it could go unseen for a time, then the SB- packages
will, during cross-compile, not inherit any one of the lookalike symbols.
You must qualify with CL: or SB-XC: depending on the need.
For example we had a related bug in MAYBE-EMIT-MAKE-LOAD-FORMS where
it should have called SB-XC:TYPEP but accidentally called CL:TYPEP.
Making the inherited TYPEP not fboundp could have caught that.