Remove thread-name deftype
Using that to restrict names of things that aren't threads (3/4ths of all
the uses) was needlessly distracting. Use SIMPLE-STRING for thread and mutex,
and be less picky about waitqueue and semaphore names - just STRING is good.

Incidentally, waitqueue-name, mutex-name, and thread-name are all SETFable
but semaphore-name isn't. I think that semaphore-name, as the outlier, is wrong.