Simplify the expansion of DEFGLOBAL
The motivation for this proved to be wrong, but the change is still nice.
I thought it might let genesis perform defglobals, which would allow removal
of !DEFINE-L-T-G, but doing so caused me an increasingly infuriating series of
crashes. Being explicit with globals that need early values is good enough.