Make sb-c::process-toplevel-form identical for +/- sb-xc-host
If it takes a screenful of comments to explain why something has to be done
differently from the right way, maybe it's time to rethink the assumptions.

With particular reference to "We have to demote COMPILE-TIME-TOO to NIL",
why is that correct? Surely there must be some subtle way in which is isn't,
or at least it's very hard to see that it's good enough. Or how about
"... tend to EVAL subforms more than once" - would it merely "tend" to or
"in fact"? Well, no matter, the change makes us follow the technically correct
compilation flow in all cases, emulating the target machine for %DEFUN etc
which is logically what genesis does as well. Bury that logic into EVAL-TLF.

When we must macroexpand twice (the one and only one such usage being in
bit-bash as of now), then clearly spell out where, what, and why.