Consistently use SB-XC: prefix on standard constants
I plan to remove the heuristic in find-free-var that checks whether a constant
matches the host's.  It's going to always be an error to use any symbol that
names a CL constant. Enforcing that found several places in code compiled in
make-host-1 which failed to use the proper symbol and got the host's value.
The change only fixes the symbols and doesn't change the leniency.