unbreak sbcl.mk
include memcpy.h through CPPFLAGS, which isn't exported to the user
through sbcl.mk (this may be an oversight, so a different mechanism
will be needed eventually).  There should be no harm in users
compiling their own C code linking to new memcpy; it is possible to
link to both symbols simultaneously.