Improve alloc-immobile-fdefn for ELF core and fix a GC bug
Save-lisp-and-die moves FDEFNs into the varyobj subspace of immobile space
so that they can get linker symbol assigned. In a similar fashion, fdefns
created dynamically need to go in varyobj subspace so that at minimum we
can map the address to the linker symbol "lisp_jit_code" which encompasses
the trailing .5 MiB of slack space written by 'editcore'.

In doing that, I discovered that for any fdefns in varyobj space which have
a macro freshly assigned (i.e. not existing before the save-lisp-and-die),
we didn't update the GC card dirty bits because INSTALL-GUARD-FUNCTION used
SAP operations to write the slots. It needs to use an asm routine.