Turn package-%local-nicknames into a weak list
We don't directly have weak lists, but they can be simulated.
Eliminate the %locally-nicknamed-by slot.  As far as I could tell,
we needed the inverted lookup as a detail of implementation - to deal with
DELETE-PACKAGE - but then exposed and documented it, which doesn't seem great.
Instead we can lazily remove obsolete nicknames if the pointee disappears.

{ADD,REMOVE}-PACKAGE-LOCAL-NICKNAME no longer need the package graph lock.
Since the mapping is not bidrectional, it suffices to use atomic-push on the
forward map. We used to grab a lock while manipulating the lists, though we
didn't check atomically for uniqueness of strings that are inserted (and still
don't) so the race that was present is still present to an extent.

This makes PACKAGE-LOCALLY-NICKNAMED-BY-LIST correct, fixes a race in
CACHED-FIND-PACKAGE, and incidentally a typo in the KEEP-OLD restart.