Fragment immobile space less badly when compiling
Cause MACROLET-DEFINITIONIZE-FUN to always use dynamic space, never
immobile space. Also if compiling an inline lambda that captured a macrolet,
do the same.  The danger of not doing the latter is that by using
EVAL-IN-LEXENV when *compile-to-memory-space* is :immobile, then you again
fragment immobile space with code that is soon to be garbage.

Minor cleanup - fold what was compile-in-lexenv into the body of CL:COMPILE
and rename what was formerly ACTUALLY-COMPILE to the new COMPILE-IN-LEXENV.
Add an extra argument to that.