Export SB-KERNEL:INTERPRETED-FUNCTION
This presents a slightly more unified interface to the interpreters, and also
causes COMPILED-FUNCTION-P not to call INTERPRETED-FUNCTION-P but instead
directly reference the layout of the instance type, whichever it is.

Unrelated: fix two tests that tried to set *evaluator-mode* to :INTERPRET
when neither interpreter was built in, which causes a type-error.