Define surrogate INSTANCE type more cautiously in XC
Restrict to the types rooted at STRUCTURE!OBJECT instead of using a union of
whatever the host defines as (OR CONDITION STRUCTURE-OBJECT STANDARD-OBJECT)
in case a host has a random thing that is one of those that we don't expect.

Also, I am mystified by the fact that returning either T or NIL for PACKAGE
objects is ok. Because in all of {ABCL, CCL, CLISP, and ECL} we get:
  (typep (find-package "CL") '(or structure-object standard-object)) => NIL
so I can not explain how that ever worked.