Run each pure test file in a randomly named package
and delete it when done, to avoid leakage from one to the next.

More of our tests can thusly be considered 'pure'. The distinction is largely
bogus anyway, as things like "filesys.pure" commit two kinds of sins:
assuming incoming state (such as claiming to know the truename of the test file
itself, which fails in a symlink tree), and assuming that calling stuff like
ENSURE-DIRECTORIES-EXIST doesn't alter state for other tests.

It's really up to the test runner to perform isolation, though conceivably
we could have a category of really dangerous tests such as the ones that
exercise the infinite-error-protector. Those are reasonably deemed "impure".
But most of our "impure" tests just want to define a few helpful DEFUNs.