Implement cross-compiler floating-point math abstraction layer
as proposed in https://sourceforge.net/p/sbcl/mailman/message/36607624/.

The most important takeaway is that floating-point literals must be spelled
in source with a leading dollar sign. Regarding other aspects of the change:

* src/code/unportable-float is gone

* a few transforms that were conditional on #+/-sb-xc-host are enabled
  - constant-folding of MAKE-{single,double}-FLOAT
  - transform of <,>,= of rational vs constant float

* inner-coerce-real-bound and inner-coerce-float-bound can directly use
  the most-{positive,negative} constants

Cleanup of more code containing comments about cross-float-infinity-kludge
remains to be done in a subsequent series of changes.

Tested with crossbuild-runner and a few alternate choices of host lisp.