Simplify dumping instances of CTYPE subtypes during bootstrap
Except for CLASSOID, treat CTYPE as a dumpable leaf, similar to how LAYOUT is
externalized. Ensure that the final image has no MAKE-LOAD-FORM for arbitrary
ctypes. I suspect it was never necessary for the compiler to dump non-classoid
type literals post-bootstrap, and it seems like a bug if it ever did that.

This change has the slightly unsatisying aspect of defining CLASSOID's
make-load-form twice, but it is less rube-goldberg-esque on the whole.