Enforce still more purity of pure tests
Disallow IN-PACKAGE which was an all-to-easy way to escape from jail.
And due to a typo, DEFMETHOD was accidentally allowed.

Rename two test files from impure to pure, and move one LOAD test from a
pure file into an impure file since it sneakily loaded a "data" file
containing an IN-PACKAGE followed by a DEFCLASS.