Fix coerce-to-condition, broken over a year ago by rev 65e0a1dd52
For want of an AND where it needed to be, we allowed any object at all
in lieu of a condition-designator. Fixing this caught a problem with
the change to fop-layout in 1a97815c90 though the root cause is deeper-
src/code/condition-boot does not set the proper bit in CONDITION,
and we were magically setting it in genesis. So do that still.

Unrelated: make CLASSOID a mandatory position arg to MAKE-LAYOUT.