rearrange vm.lisp again
This mess of circular dependencies where everything wants to be
evaluated at compile-time (and so everything it depends on needs to be
ready at read-time, ew) is horrible.  But it needs to be this way, at
least for now, so that we can give some locations to the various
register storage classes.

We now get beyond packing, and fail to emit a block header, and
subsequently fail to support type tests.  Some more platform-specific
development looms in our future...

committer: Charles Zhang <charleszhang99@yahoo.com>