define location-print-name function
it's not necessary for cross-compilation as such, but I'm spending a
lot of time in the debugger, and the debugger wants to print things,
so at least give the system a chance to print the things.

also support constants in LOCATION-PRINT-NAME, because a named
call (such as to %CHECK-BOUND) needs a constant location.  The fact
that we shouldn't have a named call to %CHECK-BOUND is, well, a
symptom of our incremental backend development

committer: Charles Zhang <charleszhang99@yahoo.com>