Implement a -C form of full reffers
As with MIPS, RISC-V has a register+immediate addressing mode; unlike
ARM, it does not have a register+register mode.  This means that if
our immediate doesn't fit in to the LW instruction's immediate field,
we can't use it (and have to compute using an interior pointer
instead).

(Computing the bounds for load/store-index is slightly involved, and
I'm not 100% sure that I've got them right.)

committer: Charles Zhang <charleszhang99@yahoo.com>