implement DEFINE-FULL-REFFER
In order to address an arbitrary word, we need an interior pointer (a
register which points, temporarily, into the middle of an object).
This needs its own storage class because, if a garbage collection
happens between the two instructions and the object is moved, the
interior pointer also needs to be updated.

committer: Charles Zhang <charleszhang99@yahoo.com>