Remove sb-c::*dump-only-valid-structures*
Binding to NIL completely obscured several things:

* debug-name markers were never dumped using their MAKE-LOAD-FORM method.

* that a user-defined MAKE-LOAD-FORM methods which returned unrepeatable
  output might accidentally behave as if it used make-load-form-saving-slots.

To the first point, compiled-debug-info can contain debug markers, but none of
the structure rooted at debug info ever gets examined to see whether it warrants
load form methods. This was due to the binding of the *dump-only* control to NIL
in DUMP-CODE-OBJECT. To deal with this, I introduced new fops for the markers,
made them dumpable leaves, and explicitly enumerated the structure-objects
that can be dumped without an entry in the validated structure table.
The make-load-form method is removed since it achieved nothing.

To the second point, the FIXME comment in FOPCOMPILE-CONSTANT is relevant.
We see instances that were not present in the validated structure table,
and for which it is no longer possible to call make-load-form.
The comment above DUMP-STRUCTURE explains why.

Also, remove obsolete reference to the :IGNORE-IT keyword.