Remove type decl contravened by EXPLICIT-CHECK
Logically they should be meaningful together, but in fact the declaration
causes the compiler to insert checks when we want the opposite - elision of
implicit checks. We end up with both an implicit and explicit type-check.
I should detect this and make it a warning.

A somewhat interesting point to note is that prior to this change
the error message for (ASH #\a 'wat) would have been:
  "The value #\a is not of type INTEGER when binding INTEGER"
but after the change it is:
  "#\a fell through ETYPECASE expression. Wanted one of (FIXNUM BIGNUM)."
It sure would be nice if we could make the message string independent
of whether the type-check was implicit or explicit.