Win32: better thread->control_stack_start calculation.
It was subtracting a page, ostensibly for the OS stack guard, but when
attaching to a foreign thread from a callback with a small stack there
is nothing left after subtraction.
Not exactly sure if it's needed, but control stack exhaustion handling
doesn't work on x86-64 either way. And other than that
control_stack_start is only used for assertions.

Fixes lp#1827946