Fix editcore some more
It remained broken after an incomplete fix for the change to compiled-debug-fun
representation. It only worked if target core pointers were "accidentally"
correct in the lisp performing the editcore operation.

As a complete surprise to me, it was not further broken by the change to
simple-fun headers. I thought I had baked in some assumptions about 6 words
per simple-fun, but apparently not.