Further reduce hash-table memory use for 64 bit machines
Store only 31 bits per precomputed hash.  This also speeds rehash
because we don't post-mix the precomputed hashes during rehash,
but instead do it when storing in the first place.

In theory we could use 32 bits per hash but drop 1 bin from the space
of 2^32 bins so as not to conflict with magic-hash-vector-value,
but I stand by the claim that having such large tables is probably
not the best idea.