Quickly skip over junk in conservative stack scan
This is mainly a consideration if we would have incurred two function calls per
word, once for preserve_pointer() and then immobile_space_preserve_pointer()
which itself tests both immobile subspaces. It wasn't so bad if immobile space
doesn't exist, but it's certainly not worse to do a simple pre-test.