Add a PATHNAMEP bit to LAYOUT
PATHNAMEP shows up the implementation of standard predicates due to
language requirements for EQUAL, EQUALP, and the respective hash functions.
Since most instances are not pathnames, but most are at the depthoid of
pathname or greater, it tooks several tests to rule out pathnamep on most
instances. This change reduces the test to a simpler one.

And as we already had a bit for structure-object, use that too.