Do something useful with very small values of rehash-size
Use the behavior before the de-wonkification change if the resize amount
is anything other than the default because I don't care to think about it.

I suspect that prior to my previous change, we caused positive growth on each
resize due to the use of "size+1" which was obviously intended to be used for
computing the sizes of the vectors allowing for the unused leading cell,
but it sort of accidentally did the right thing to enforce sizing up as well.