Give REMHASH the same treatment as PUTHASH
Add 5 specializations, 4 standard plus a user-defined predicate.

Fix a problem with consing in PUTHASH on 32-bit machines. Because we lack an
unboxed full call convention, it is not fesible to pass (UNSIGNED-BYTE 32) hash
values without boxing. Thus the helper defuns utilized by {PUT,REM}HASH can't be
toplevel. A toplevel FLET enclosing the defuns is possible, which has its own
problem: genesis is unable to bind non-toplevel functions to their names, so the
family of functions is inaccessible to MAKE-HASH-TABLE until too late.
But it _does_ work (a little surprisingly) to write a toplevel defun containing
an FLET that returns lambdas that we bind to toplevel names via setf.
Hence, the obfuscated idiom employed in this change.