sb-bsd-sockets: Somewhat fix (SETF SOCKOPT-BIND-TO-DEVICE)
The socket option value is a string which DEFINE-SOCKET-OPTION
translates to

  (sb-alien:with-alien ((buffer sb-alien:c-string))
     (setf buffer ...)
     (sb-bsd-sockets-internal::setsockopt ... buffer ...))

Passing BUFFER as an argument like this invokes the "naturalization"
so that a lisp string is passed to SETSOCKOPT which wants an
(alien (* t)). This resulted in a compile-time warning when building
the contrib and a runtime type-error when calling the function.

The solution is more of a workaround:
1) take the address of the local alien
2) cast it to (* (* t))
3) dereference it yielding an (alien (* t))
This avoids "naturalization" and matches the expected alien type.

strace indicates that everything works, but the system call seems to
require elevated privileges, so no test.