Make allocate-static-vector threadsafe
Remove the store at one word past the new object's end which might have had
something to do with cheneygc, but even so, it seemed to serve no purpose.

Correct the boundary condition- it should be '<=' static-space-end
but it had to be strictly '<' due to the store beyond the allocation.