Fix special unbinding when unwinding is being itself unwound on x86-64.
Need to perform unbinding before unlinking the UWP so that if
interrupted while unbinding it can still run the clean up form. While
the cleanup form itself cannot be protected from interrupts (can't run
it twice) one of the variables being unbound can be
*interrupts-enabled*.

Other backends still use the old way of unbinding in the clean up form
and will have to transition to this scheme.