Reduce insanity level of ppc assembler
Make DEF-PPC-IFORMAT a global macro. Localize other macros to point of use.

Reject ZERO-TN as RA operand in any instructions that do not allow encoding it
as a register. Require instead that the instruction receive literal 0, which
is what the encoding of 0 means. This liberates the assembler from our specific
use of r0 in lisp code which will free up a register.
Literally the only reason for forcing r0 to hold zero (aside from minor
simplifications in codegen where 0 is stored to memory) was that the assembler
did not know whether we meant to use a register or constant, and could not
signal an error in contexts where the r0 was used as an operand but would not
do the right thing if it held other than zero. But we _can_ signal an error
at assembly time (The fact that IBM native syntax is superficially ambiguous
is beside the point. It's not actually ambiguous, but merely unclear)

Group all the load/store instructions together and more closely
in the order of appearance in the architecture reference manual.

Eventually I plan to merge all instructions into one file,
but for the time being all these changes are made in two places.