ppc64: avoid wiring a register to zero
It seems to have become tradition to define a storage class for the constant 0,
which wastes a register on machines that don't have a register that always
reads as zero.
This change uses r0 as the temp register for load/store via tagged pointers
which frees up non-descriptor reg NL6 for its intended use.