Show code page count in print_generation_stats
When making this change, I noticed that the WP count in gen0 is often nonzero.
I didn't think we allowed that. It _should_ make sense because we could track
written pages of nursery generation large vectors, to reduce scavenging in
newspace. But we don't do that, so ... WHAT IS GOING ON? I can't even.