Re-allow choice of n-fixnum-tag-bits for #+64-bit
We need it for ppc64 so that instance-pointer-lowtag can end in #b00
which should eliminate an extra 'li' for instance-ref/set.
I'm starting with unbreaking x86-64 first, since it used to work
but stopped working due to assumptions about the range of fixnums
in some architecture-independent and architecture-dependent code.

(I don't claim that rearranging lowtags will actually succeed)