Improve error message if small spaces can't be mapped
Depending on the architecture, add the MAP_32BIT flag to fixed mappings which
do not - on Linux at least - pass the MAP_FIXED flag to mmap().
Essentially, if we don't get the address we need, then mmap() should give us
an address that might work (contingent on total rebuild of SBCL and user code)
for the small spaces instead of an address that can't possibly work.