consistency check for collation hash table
On 64-bit, we use the fact that three codepoints fit in a 63-bit
fixnum and make the hash table use EQ as the test, rather than EQL.
But because our fixnums might be smaller (currently 61 bits on ppc64,
for example), this isn't necessarily a valid optimization.  However,
the high codepoints in Unicode are private use / unallocated, and do
not currently show up in any collation tables; add an averral to
enforce this at build-time.