remove a Unicode not-constant
The number of collation entries varies between versions of Unicode.
It is useful to check that our binary database has the right number of
entries, but generate that number of entries to check against from the
initial parse of the Unicode data file, rather than hard-coding the
value in SBCL sources.