Remove indirection from load_core_bytes to os_map
os_map was badly named, as two variations boiled down to read(), not mmap().
Rename all of them to load_core_bytes() which was the only consumer.

Fix the failure case not to call perror() unless the system call fails
and not to call lose() either.

Remove the extra junk in the HPUX code - there is apparently no need to loop
nor call lseek to obtain the incoming file position which was not later used.