Enhance disassembler to optionally produce instruction model
If STREAM is NIL in MAP-SEGMENT-INSTRUCTIONS, the instruction printer will push
a representation of each operand into DSTATE-OPERANDS rather than output to the
stream. To utilize this for any given backend, all its printers must be made
aware of the NIL stream convention.

This change takes a worse-is-better approach. The "elegant" technique would have
been to always convert to the abstract model first, and then optionally produce
textual output from the model or not. But the models as they are need not be
perfectly correct renderings of the decoding - they merely have to be adequate
for certain kinds of examination of assembly code.