Use FUN_RAW_ADDR_OFFSET in fewer places
This constant must be utilized with care because of the differing ways in
which we compute a delta from a tagged pointer to entry point: in the
'raw_addr' slot of an fdefn vs the 'self' slot of a simple-fun or closure.

The choices are binary, yielding 4 theoretical possibilities:
 * tagged   fdefn->raw_addr, tagged   simple_fun->self (sparc,arm,riscv)
 * untagged fdefn->raw_addr, untagged simple_fun->self (x86oid)
 * untagged fdefn->raw_addr, tagged   simple_fun->self (all others)
 * tagged   fdefn->raw_addr, untagged simple_fun->self (NOT USED)

As defined, the FUN_RAW_ADDR_OFFSET is correct for the first two cases
in either context (fdefn or function), but correct for the third case only
with respect to fdefns, and not with functions.

Inspection of each point of use turned up the following:
* in coreparse, FUN_RAW_ADDR_OFFSET was misused, but in code that was
  not compiled in unless DEBUG_CORE_LOADING was defined.

* in gc-common, its use on a simple-fun was guarded by X86|X86_64,
  and the '#else' branch wired in the constant 0 which is right;
  its use with a closure was guarded by X86|X86_64.

* in fullcgc, its use with a closure was guarded by X86|X86_64.

* in traceroot it was guarded by FUN_SELF_FIXNUM_TAGGED which is
  equivalent to X86|X86_64.

After this change, it is likely possible to remove some #if guards.