remove memcpy glibc workaround
While it would be nice to continue to build sbcl binaries that work on
older Linuxes from newer ones, the bug fix for fcntl() in glibc-2.28
means that doing so would not just be a slight performance decrease,
but including by default a buggy library call rather than a fixed
one.  SBCL continues to support those older releases, but binaries for
them should be produced with earlier toolchains.