Untangle fixups from EAs in x86-64 assembler
Amongst the reasons our 32-bit x86 assembler is problematic is that most
uses of fixups treat them implicitly as EAs. This notion was copied into
the 64-bit assembler. But a fixup is not an EA, it is merely a constant
plugged in when a fasl is loaded. It could be the displacements of an EA,
but in general it can be any imm32 operand in any instruction.
Imparting rigor to this distinction - that is, explictly constructing an
EA when you need it - is a step toward making the assembler more rational.

Given that the 64-bit code was already much less bad (in this regard) than
the 32-bit, this change rips out smelly parts that were blindly borrowed.

Additionally, the :ALLOW-CONSTANTS argument to EMIT-EA is unnecessary
in the 64-bit assembler, and probably also in the 32-bit assembler,
but I suspect that it has to do with conflation of EAs and fixups.